update fixtures + gift routes

+++ src/Controller/GiftController.php
@@ -5,9 +5,10 @@ namespace App\Controller;
 use App\Entity\Gift;
 use App\Entity\User;
 use Doctrine\ORM\EntityManagerInterface;
+use Doctrine\ORM\Exception\ORMException;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\BrowserKit\Request;
 use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
@@ -16,13 +17,40 @@ use Symfony\Component\Serializer\SerializerInterface;
 class GiftController extends AbstractController
 {
 
+    /**
+     * @throws ORMException
+     */
     #[Route('/api/gift/{id}', name: 'deleteGift', methods: ['DELETE'])]
     public function deleteGift(Gift $gift, EntityManagerInterface $em): JsonResponse
     {
-        $em->remove($gift);
-        $em->flush();
+        /** @var User $user */
+        $user = $this->getUser();
+        $userGiftList = $user->getGiftList();
 
+        if (!$userGiftList->getGifts()->contains($gift)) {
+            return new JsonResponse('Vous n\'avez pas les droits pour supprimer ce cadeau', Response::HTTP_FORBIDDEN);
+        }
+
+        try {
+            $em->remove($gift);
+            $em->flush();
+        } catch (ORMException $e) {
+            throw new ORMException('erreur : ' . $e);
+        }
         return new JsonResponse(null, Response::HTTP_NO_CONTENT);
+    }
+
+    #[Route('/api/gifts', name: "listGift", methods: ['GET'])]
+    public function getGift(SerializerInterface $serializer): JsonResponse
+    {
+
+        /** @var User $user */
+        $user = $this->getUser();
+        $giftlist = $user->getGiftList();
+
+        $jsonGiftlist = $serializer->serialize($giftlist, 'json', ['groups' => 'getGifts']);
+
+        return new JsonResponse($jsonGiftlist, Response::HTTP_OK, [], true);
     }
 
     #[Route('/api/gift', name: "createGift", methods: ['POST'])]
@@ -45,7 +73,7 @@ class GiftController extends AbstractController
         $jsonGiftlist = $serializer->serialize($gift, 'json', ['groups' => 'getGift']);
 
         $location = $urlGenerator->generate(
-            'detailBook',
+            'listGift',
             ['id' => $giftlist->getId()],
             UrlGeneratorInterface::ABSOLUTE_URL
         );
